# Conflicts:
#	core/assets/metadata/checksums.json
#	core/assets/tilesets/diffuse-entities.atlas
#	core/assets/tilesets/normal-entities.atlas
#	core/src/main/java/technology/rocketjump/saul/rendering/GameRenderer.java
#	core/src/main/java/technology/rocketjump/saul/ui/SelectableOutlineRenderer.java

+++ core/src/main/java/technology/rocketjump/saul/entities/ai/combat/CreatureCombat.java
@@ -63,7 +63,7 @@ public class CreatureCombat {
 	}
 
 	public int getWeaponRangeAsInt() {
-		return (int)Math.max(1, equippedWeapon.getRange());
+		return Math.max(1, equippedWeapon.getRange());
 	}
 
 	public int maxDefensePool() {
